requirements.txt
attrs==23.1.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.6.2

Brotli==1.0.9
certifi==2022.9.24
chardet==5.1.0
click==8.1.6

colorama==0.4.6

configobj==5.0.8
cryptography==38.0.4

dbus-python==1.3.2

distro==1.8.0

duplicity==1.2.2
fasteners==0.17.3
freetype-py==2.4.0
future==0.18.2

gpg==1.10.0
greenlet==2.0.2

httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.2.0

launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
libevdev==0.5
lockfile==0.12.2
louis==1.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
msgpack==1.0.3
mutagen==1.46.0
natsort==8.0.2
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
packaging==23.1
paramiko==2.12.0
pexpect==4.8.0
Pillow==10.0.0
ptyprocess==0.7.0
pycairo==1.24.0
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.46.0
PyJWT==2.7.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==3.1.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1

python-dateutil==2.8.2
python-debian==0.1.49
python-uinput==0.11.2
pytz==2023.3
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.1
reportlab==4.0.4
requests==2.31.0
rich==13.3.1
rlPyCairo==0.3.0
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.5
systemd-python==235


ubuntu-pro-client==31.2.4


urllib3==1.26.16

wadllib==1.3.6
websockets==10.4
wxPython==4.2.1
xdg==5

xyzservices==2023.2.0
yt-dlp==2022.4.8
zipp==1.0.0

fastapi==0.115.2
pydantic==2.9.2
uvicorn==0.31.1
aiohttp==3.10.10
SQLAlchemy==2.0.35
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1